Complete overhaul
Haven't touched this project in almost a year, recently I became interested in cellular automata and decided to pick it up again. Since it was my first real project in C, the code was bad, and I wanted to use C++ this time, so I rewrote everything from the ground up. The new system should be much more expandable and also run better in general. Many more changes to come soon.